Also define NLGA when we build with atomic ops (#80)
This uncovered the use of NLGA to guard the initializaition
of the valid mem cache. This code is removed, as it isn't
working properly for a per-thread cache anyways. What's more,
it shouldn't be required anyways since static data is guaranteed
to be initialized to zero anyways.

Fixes https://github.com/libunwind/libunwind/issues/79