dwarf: allocate a debug frame index table with mmap()
By analogy with "dwarf: do not allocate in load_debug_frame (#72)"
avoid to call realloc(). An index table is also created infrequently
and freed in unw_flush_cache() only.

The main idea of the patch is to calculate an index table size first
and then allocate a whole memory region.